requirements.txt
asgiref==3.11.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
inflection==0.5.1
Markdown==3.10.2
packaging==26.1
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0